# nest-backend/openApiSpec.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/workspaces/{workspaceId}/projects": {
            "post": {
                "operationId": "ProjectController_create",
                "parameters": [
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProjectDto"
                            }
                        }
                    }
                },
                "responses": { "201": { "description": "" } },
                "tags": ["Projects"]
            },
            "get": {
                "operationId": "ProjectController_findMany",
                "parameters": [
                    {
                        "name": "name",
                        "required": false,
                        "in": "query",
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": { "200": { "description": "" } },
                "tags": ["Projects"]
            }
        },
        "/workspaces/{workspaceId}/projects/{projectId}": {
            "get": {
                "operationId": "ProjectController_findOne",
                "parameters": [
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "projectId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": { "200": { "description": "" } },
                "tags": ["Projects"]
            },
            "delete": {
                "operationId": "ProjectController_delete",
                "parameters": [
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "projectId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": { "200": { "description": "" } },
                "tags": ["Projects"]
            }
        },
        "/workspaces/{workspaceId}/projects/{id}": {
            "patch": {
                "operationId": "ProjectController_update",
                "parameters": [
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProjectDto"
                            }
                        }
                    }
                },
                "responses": { "200": { "description": "" } },
                "tags": ["Projects"]
            }
        },
        "/workspaces/{workspaceId}/projects/{id}/uploads": {
            "patch": {
                "operationId": "ProjectController_uploadFile",
                "parameters": [
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    }
                },
                "tags": ["Projects"]
            }
        },
        "/workspaces/{workspaceId}/projects/{id}/uploads/{uploadId}": {
            "patch": {
                "operationId": "ProjectController_deleteUpload",
                "parameters": [
                    {
                        "name": "workspaceId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "uploadId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    }
                },
                "tags": ["Projects"]
            }
        },
...
        "/tasks": {
            "post": {
                "operationId": "create_task",
                "summary": "",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTaskDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                },
                "tags": ["Tasks"]
            }
        },
        "/tasks/{id}": {
            "get": {
                "operationId": "find_one_task",
                "summary": "",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                },
                "tags": ["Tasks"]
            },
            "patch": {
                "operationId": "update_task",
                "summary": "",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskDto"
                            }
                        }
                    }
                },
                "responses": { "200": { "description": "" } },
                "tags": ["Tasks"]
            },
            "delete": {
                "operationId": "delete_task",
                "summary": "",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": { "204": { "description": "" } },
                "tags": ["Tasks"]
            }
        },
        "/tasks/{id}/uploads": {
            "patch": {
                "operationId": "create_task_upload",
                "summary": "",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                },
                "tags": ["Tasks"]
            }
        },
        "/tasks/{taskId}/uploads/{uploadId}": {
            "delete": {
                "operationId": "delete_task_upload",
                "summary": "",
                "parameters": [
                    {
                        "name": "taskId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    },
                    {
                        "name": "uploadId",
                        "required": true,
                        "in": "path",
                        "schema": { "type": "number" }
                    }
                ],
                "responses": { "204": { "description": "" } },
                "tags": ["Tasks"]
            }
        },
...
    },
    "info": {
        "title": "Posters example",
        "description": "The posters API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "Upload": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "fieldname": { "type": "string" },
                    "originalname": { "type": "string" },
                    "encoding": { "type": "string" },
                    "contentEncoding": { "type": "string" },
                    "mimetype": { "type": "string" },
                    "size": { "type": "number" },
                    "bucket": { "type": "string" },
                    "key": { "type": "string" },
                    "acl": { "type": "string" },
                    "contentType": { "type": "string" },
                    "contentDisposition": { "type": "string" },
                    "storageClass": { "type": "string" },
                    "serverSideEncryption": { "type": "string" },
                    "metadata": { "type": "object" },
                    "location": { "type": "string" },
                    "etag": { "type": "string" },
                    "versionId": { "type": "string" }
                },
                "required": [
                    "id",
                    "fieldname",
                    "originalname",
                    "encoding",
                    "contentEncoding",
                    "mimetype",
                    "size",
                    "bucket",
                    "key",
                    "acl",
                    "contentType",
                    "contentDisposition",
                    "storageClass",
                    "serverSideEncryption",
                    "metadata",
                    "location",
                    "etag",
                    "versionId"
                ]
            },
...
            "CreateProjectDto": {
                "type": "object",
                "properties": {
                    "name": { "type": "string" },
                    "workspaceId": { "type": "number" },
                    "tenantId": { "type": "number" }
                },
                "required": ["name", "workspaceId", "tenantId"]
            },
            "UpdateInformationTabDto": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "description": { "type": "string" },
                    "topics": {
                        "type": "array",
                        "items": { "type": "object" }
                    },
                    "keywords": {
                        "type": "array",
                        "items": { "type": "object" }
                    },
                    "attachments": {
                        "type": "array",
                        "items": { "type": "object" }
                    }
                }
            },
            "UpdateProjectDto": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "name": { "type": "string" },
                    "visibility": { "type": "object" },
                    "informationTab": {
                        "$ref": "#/components/schemas/UpdateInformationTabDto"
                    },
                    "tenantId": { "type": "number" }
                },
                "required": [
                    "id",
                    "name",
                    "visibility",
                    "informationTab",
                    "tenantId"
                ]
            },
...
            "User": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "objectId": { "type": "string" },
                    "organizationName": { "type": "string", "nullable": true },
                    "firstName": { "type": "string" },
                    "lastName": { "type": "string" },
                    "email": { "type": "string" },
                    "isDummyUser": { "type": "boolean" },
                    "subscribed": { "type": "boolean" },
                    "privacyPolicy": { "type": "boolean" },
                    "termOfUse": { "type": "boolean" },
                    "disableNotifications": { "type": "boolean" },
                    "linkedin": { "type": "string" },
                    "researchGate": { "type": "string" },
                    "twitter": { "type": "string" },
                    "currentPosition": { "type": "string", "nullable": true },
                    "biography": { "type": "string", "nullable": true },
                    "createdAt": { "format": "date-time", "type": "string" },
                    "updatedAt": { "format": "date-time", "type": "string" },
                    "avatarId": { "type": "number", "nullable": true },
                    "avatar": { "$ref": "#/components/schemas/Upload" },
                    "posters": {
                        "type": "array",
                        "items": { "type": "object" }
                    },
                    "authoredPosters": {
                        "type": "array",
                        "items": { "type": "object" }
                    },
                    "roles": { "type": "array", "items": { "type": "object" } },
                    "comments": {
                        "type": "array",
                        "items": { "type": "object" }
                    },
                    "userDirectory": { "type": "object" },
                    "skills": { "type": "array", "items": { "type": "object" } }
                },
                "required": [
                    "id",
                    "objectId",
                    "organizationName",
                    "firstName",
                    "lastName",
                    "email",
                    "isDummyUser",
                    "subscribed",
                    "privacyPolicy",
                    "termOfUse",
                    "disableNotifications",
                    "linkedin",
                    "researchGate",
                    "twitter",
                    "currentPosition",
                    "biography",
                    "createdAt",
                    "updatedAt",
                    "avatarId"
                ]
            },
...
            "Status": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "name": { "type": "string" },
                    "boardId": { "type": "number" },
                    "createdAt": { "format": "date-time", "type": "string" },
                    "updatedAt": { "format": "date-time", "type": "string" }
                },
                "required": ["id", "name", "boardId", "createdAt", "updatedAt"]
            },
            "Decimal": { "type": "object", "properties": {} },
            "Task": {
                "type": "object",
                "properties": {
                    "id": { "type": "number" },
                    "title": { "type": "string" },
                    "description": { "type": "string" },
                    "status": { "$ref": "#/components/schemas/Status" },
                    "positionByStatus": {
                        "$ref": "#/components/schemas/Decimal"
                    },
                    "attachments": {
                        "type": "array",
                        "items": { "type": "object" }
                    },
                    "assignees": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/User" }
                    },
                    "dueDate": { "format": "date-time", "type": "string" },
                    "createdAt": { "format": "date-time", "type": "string" },
                    "updatedAt": { "format": "date-time", "type": "string" }
                },
                "required": [
                    "id",
                    "title",
                    "description",
                    "status",
                    "positionByStatus",
                    "attachments",
                    "assignees",
                    "dueDate",
                    "createdAt",
                    "updatedAt"
                ]
            },
...
            "CreateTaskDtoProject": {
                "type": "object",
                "properties": {
                    "visibility": {
                        "type": "string",
                        "enum": ["PUBLIC", "PRIVATE"]
                    },
                    "workspaceId": { "type": "number" }
                },
                "required": ["visibility", "workspaceId"]
            },
            "CreateTaskDtoBoard": {
                "type": "object",
                "properties": {
                    "project": {
                        "$ref": "#/components/schemas/CreateTaskDtoProject"
                    },
                    "tenantId": { "type": "number" },
                    "id": { "type": "number" }
                },
                "required": ["project", "tenantId", "id"]
            },
            "CreateTaskDto": {
                "type": "object",
                "properties": {
                    "title": { "type": "string" },
                    "description": { "type": "string" },
                    "statusId": { "type": "number" },
                    "board": {
                        "$ref": "#/components/schemas/CreateTaskDtoBoard"
                    }
                },
                "required": ["title", "description", "statusId", "board"]
            },
            "AssigneeDto": {
                "type": "object",
                "properties": { "id": { "type": "number" } },
                "required": ["id"]
            },
            "UpdateTaskDto": {
                "type": "object",
                "properties": {
                    "title": { "type": "string" },
                    "description": { "type": "string" },
                    "statusId": { "type": "number" },
                    "positionByStatus": { "type": "number" },
                    "dueDate": { "format": "date-time", "type": "string" },
                    "assignees": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/AssigneeDto" }
                    }
                }
            },
...
        }
    }
}